So now I remember why I had the SafeEnumerable class crap - it wasn't thread safety, it was collection modification safety! I remember running into this issue months back and realizing that it is going to be a design nightmare to fix in a way that is easy to use and safe.
The problem is that we have this:
- World.Update() ... foreach (var map in Maps) map.Update()
-- Map.Update() ... foreach (var entity in Entities) entity.Update()

and this:
- HandlePackets ... foreach (var recvPacket in ReceivedPackets) recvPacket.Process()

Now, the problem here is that we end up with crap like this:
- World.Update()
-- Map.Update()
--- Character.Update()
---- Character.Teleport(), Character.Kill(), Character.Attack() (leading to Character.Kill() somewhere else), etc

All of those Character methods will change something on the map. They will move, new characters are added, or existing characters are removed. In any case, the enum from before (looping over all entities to update) will be altered.

The easy hack to fix this would be to just to .ToArray() at the end of the enum before enumerating. I don't care if we use the "old state" of the enum. All of the conditions, such as Update() or Attack() or whatever, should be checking if the values are valid anyways. But that would be difficult in we'd have to remember to do that everywhere, and my god - that'd be a shit-ton of overhead!

Honestly, I am not sure what to do right now...