AutomaticAnimatedGrhData is in! Bwahahahaha!
Basically, things will work just like before. Though instead, the frames of the automatically-added animated GrhDatas will not be added as StationaryGrhData. This way, the frames are determined at runtime. This leads to the ability to automatically update all the frames and even the speed without any extra hassle, and the AutomaticAnimatedGrhData only requires the GrhIndex and SpriteCategorization in the grhdata.xml file.