Finally fixed the background images to render in the correct order. Changed IDrawable's LayerDepth value from Byte to Int. Having negative values will be handy to push some back (namely the MapGrh). This has no storage impact on anything but the MapGrh, which internally forces in the range of a short. I highly doubt anyone needs more than 65536 layers. This was mostly done for the BackgroundImage since the LayerDepth is acquired from the Depth, which can be a huge range of numbers as it is a float.
Also added GrhIndex.Invalid, which is a unique value that cannot be recreated using new GrhIndex(). This is to provide stuff a way to say, "this GrhIndex is not valid", without having to use a nullable GrhIndex. I didn't want to make it so it was troublesome to "try" to load the invalid GrhIndex (like with nullable), just that it would never be able to happen.